+++ jjikeobang-server/src/main/java/com/jjikeobang/member/controller/MemberCheckIdController.java
@@ -1,0 +1,7 @@
+package com.jjikeobang.member.controller;
+
+import jakarta.servlet.annotation.WebServlet;
+
+@WebServlet("/member/check-id")
+public class MemberCheckIdController {
+}
