+++ jjikeobang-server/src/main/java/com/jjikeobang/member/controller/MemberLoginController.java
@@ -1,0 +1,28 @@
+package com.jjikeobang.member.controller;
+
+import com.jjikeobang.member.model.Member;
+import com.jjikeobang.member.service.MemberServiceImpl;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+import java.util.List;
+
+@WebServlet("/member/login")
+public class MemberLoginController extends HttpServlet {
+
+    private final MemberServiceImpl memberService;
+
+    public MemberLoginController() {
+        this.memberService = new MemberServiceImpl();
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+        super.doPost(req, resp);
+    }
+}
